% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusClust.R
\name{consensusClust}
\alias{consensusClust}
\title{Perform consensus clustering on a scRNA-seq matrix}
\usage{
consensusClust(
  data,
  pcNum = 15,
  nboots = 200,
  clusterFun = "leiden",
  bootSize = 0.8,
  resRange = seq.int(0.05, 1, by = 0.05),
  pcaMethod = "irbla",
  kNum = 30,
  mode = "robust",
  threads = 1,
  assay = "RNA",
  ...
)
}
\arguments{
\item{data}{matrix of scaled counts, or a Seurat or SingleCellExperiment object from which these can be extracted}

\item{pcNum}{number of principal components to consider when finding nearest neighbours}

\item{nboots}{number of boostraps to perform}

\item{clusterFun}{Which of the igraph community detection algorithms to use for clustering: "leiden" or "louvain"}

\item{bootSize}{fraction of total cell number to use per bootstrap}

\item{resRange}{Numeric vector of resolutions to cluster over}

\item{pcaMethod}{Method for computing pca (specify irbla for faster, less accurate PCA)}

\item{kNum}{number of nearest neighbours for knn graph construction}

\item{mode}{How to deal with cluster assignments from different resolutions, either pick the one with highest silhouette score ('robust')
or use all when deciding consensus clusters ('granular')}

\item{threads}{How many cpus to use in parallel}

\item{assay}{Name of seurat or SingleCellExperiment assay, if providing an object rather than scaled counts}
}
\value{
vector: character vector of consensus clustering assignments
}
\description{
Perform consensus clustering on a scRNA-seq matrix
}
