% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusClust.R
\name{determineHierachy}
\alias{determineHierachy}
\title{Conducts PCA, and clusters cells based on resulting PCs, returning assignments as a dataframe}
\usage{
determineHierachy(distanceMatrix, assignments)
}
\arguments{
\item{distanceMatrix}{Matrix of cell-cell distances}

\item{assignments}{clustering assignments}
}
\description{
Conducts PCA, and clusters cells based on resulting PCs, returning assignments as a dataframe
}
