% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusClust.R
\name{getClustAssignments}
\alias{getClustAssignments}
\title{Conducts PCA, and clusters cells based on resulting PCs, returning assignments as a dataframe}
\usage{
getClustAssignments(
  scaledCounts,
  pcNum,
  clusterFun = "leiden",
  resRange,
  pcaMethod,
  kNum,
  mode = "robust",
  cellOrder,
  ...
)
}
\arguments{
\item{scaledCounts}{Matrix of scaled counts}

\item{pcNum}{number of principal components to consider when finding nearest neighbours}

\item{clusterFun}{Which of the igraph community detection algorithms to use for clustering: "leiden" or "louvain"}

\item{resRange}{Numeric vector of resolutions to cluster over}

\item{pcaMethod}{Method for computing pca (specify irbla for faster, less accurate PCA)}

\item{kNum}{number of nearest neighbours for knn graph construction}

\item{mode}{How to deal with cluster assignments from different resolutions, either pick the one with highest silhouette score ('robust')
or use all when deciding consensus clusters ('granular')}

\item{cellOrder}{Vector of non-boostrapped cellnames in the order of the non-boostrapped data matrix}
}
\description{
Conducts PCA, and clusters cells based on resulting PCs, returning assignments as a dataframe
}
