% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusClust.R
\name{testSplits}
\alias{testSplits}
\title{Generate a null count matrix from a scDesign3 paramters, process and cluster, and return the silhouette score of the clustering}
\usage{
testSplits(
  sce,
  pca,
  dend,
  kNum,
  alpha,
  finalAssignments,
  varsToRegress,
  BPPARAM,
  silhouetteThresh,
  ...
)
}
\description{
Generate a null count matrix from a scDesign3 paramters, process and cluster, and return the silhouette score of the clustering
}
