% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusClust.R
\name{determineHierachy}
\alias{determineHierachy}
\title{Computes dendrogram between clusters}
\usage{
determineHierachy(distanceMatrix, assignments, return = "dendrogram")
}
\arguments{
\item{distanceMatrix}{Matrix of cell-cell distances}

\item{assignments}{clustering assignments}
}
\description{
Computes dendrogram between clusters
}
