% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusClust.R
\name{testSplits}
\alias{testSplits}
\title{Generate a null count matrix from a scDesign3 paramters, process and cluster, and return the silhouette score of the clustering}
\usage{
testSplits(
  sce,
  pca,
  dend,
  kNum,
  alpha,
  finalAssignments,
  varsToRegress,
  BPPARAM,
  silhouette,
  silhouetteThresh,
  test_sep = F,
  resRange = seq(0.1, 3.4, 0.15),
  ...
)
}
\description{
Generate a null count matrix from a scDesign3 paramters, process and cluster, and return the silhouette score of the clustering
}
